Move fixtures to own file, make fixture naming more consistent
